+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Initial release
+
+## 0.0.2
+
+* Fix chart height not consistent with small amount of tasks
+* Add optional `scrollWhileDrag` to enable/disable scrolling while dragging tasks on edge of screen